Add resource_id to get_metric validation schema
As nflex connector support to get metrics for multiple resource at once,
we added resource_id field to the result data of get_metrics handler so
we can identify which metric belong to which resource.

This filed can be omitted, in this case that metrics will be associated
with resource which is passed to this handler. this is the same
bahaviour as before.